+++ src/com/kh/MasterPiece/test/model/service/testService.java
@@ -1,0 +1,22 @@
+package com.kh.MasterPiece.test.model.service;
+
+
+
+import static com.kh.MasterPiece.common.JDBCTemplate.close;
+import static com.kh.MasterPiece.common.JDBCTemplate.getConnection;
+
+import java.sql.Connection;
+
+import com.kh.MasterPiece.test.model.dao.testDao;
+
+public class testService {
+	public int[] Count(){
+		Connection con = getConnection();
+		
+		int list[] = new testDao().Count(con);
+		
+		close(con);
+		
+		return list;
+	}
+}
