--- src/sql/member/member-query.properties
@@ -1,9 +1,0 @@
-
-insertMember=INSERT INTO MEMBER(USER_NO, USER_TYPE, USER_NAME, USER_ID, USER_PWD, PHONE, ADDRESS, EMAIL, STATUS)  VALUES (USER_NO.NEXTVAL, DEFAULT, ?, ?, ?, ?, ?, ?, DEFAULT)
-loginSelect=SELECT * FROM MEMBER WHERE USER_ID = ? AND USER_PWD = ?
-idCheck=SELECT COUNT(*) FROM MEMBER WHERE USER_ID = ?
-emailCheck=SELECT COUNT(*) FROM MEMBER WHERE EMAIL = ?
-
-selectCount=select t, count(t) from test group by t order by t
-insert=insert into test values(test_no.nextval, ?)
-
