--- src/sql/notice/notice-query.properties
@@ -1,4 +1,0 @@
-selectList=SELECT N.NNO, N.NTITLE, N.NCONTENT, M.NICK_NAME, N.NCOUNT, N.NDATE FROM NOTICE N JOIN MEMBER M ON(N.NWRITER = M.UNO) WHERE N.STATUS = 'Y' ORDER BY NNO DESC
-insertNotice=INSERT INTO NOTICE VALUES(SEQ_NNO.NEXTVAL, ?, ?, ?, DEFAULT, ?, DEFAULT)
-selectOne=SELECT NNO, NTITLE, NCONTENT, NICK_NAME, NCOUNT, NDATE FROM NOTICE N JOIN MEMBER M ON (N.NWRITER = M.UNO) WHERE N.STATUS = 'Y' AND NNO = ?
-updateCount=UPDATE NOTICE SET NCOUNT=(SELECT NCOUNT FROM NOTICE WHERE NNO = ?) + 1 WHERE NNO = ?
