+++ src/sql/test-query.properties
@@ -1,0 +1,1 @@
+selectCount=select BOARD_CATEGORY, count(BOARD_CATEGORY) from board group by BOARD_CATEGORY order by BOARD_CATEGORY
