--- web/views/admin/common/errorPage.jsp
@@ -1,13 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<% String msg = (String)request.getAttribute("msg"); %>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-</head>
-<body>
-	<h1 align="center"><%= msg %></h1>
-</body>
-</html>
