--- web/views/admin/common/footer.jsp
@@ -1,88 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-
-<style>
-
-</style>
-</head>
-<body>
-	<hr>
-	<p>Copyright. All rights reserved</p>
-
-
-<style>
-.footerHr { border-color: #f43641; }
-.footerP { text-align:center; }
-.footer1{ display: inline-block; }
-.info_conpany { font-size:11px; letter-spacing: -1px; line-height: 80%; border-right: #e5e5e5 1px solid; padding: 0 40px; display: inline-block;}
-.info_tel{ color:#000000; font-size:11px; display:block; line-height: 70%; margin-bottom:4px; letter-spacing:-1px; display: inline-block;}
-.footerCopyright{ margin-top: -30px;}
-.footerLogo{padding-top:-50px;}
-.footerCopyright .footerInfo .info_conpany .btn {display:inline-block; height:15px; padding:0 4px; border:1px solid #ccc; background-color:#fff; color:#999; font-size:11px; line-height:15px; margin-left:5px; text-decoration:none;}
-.footerCopyright .footerInfo .info_tel .btn{display:inline-block; height:15px; padding:0 4px; border:1px solid #ccc; background-color:#fff; color:#999; font-size:11px; line-height:15px; margin-left:5px; text-decoration:none;}
-.btnB{display:inline-block; height:15px; padding:8px 61px; border:1px solid #ccc; background-color:#eaeaec; color:#000000; font-size:14px; line-height:14px; font-family:Noto-R; text-decoration:none;}
-
-
-</style>
-
-</head>
-<body>
-	<hr class="footerHr">
-	<br>
-	
-	<div class="footerCopyright" align="center">
-		<div class="footerLogo footer1">
-			<a><img src="../../images/main/MasterPiece_logo.png" style="width:190px; height:130px;"></a>
-		</div>
-		<div class="footerInfo footer1" align="left">
-			<ul style="list-style-type:none; margin-left: -40px;">
-				<li class="info_conpany">
-					<p style="font-weight:bold; font-size:14px;">㈜ MasterPiece</p>
-					<p>서울특별시 강남구 역삼동</p>
-						<p>
-						대표이사 : MasterPiece
-						｜
-						사업자등록번호 : 012-34-56789
-						<a href="http://www.ftc.go.kr/info/bizinfo/communicationView.jsp?apv_perm_no=2001302009530200417&amp;area1=&amp;area2=&amp;currpage=1&amp;searchKey=04&amp;searchVal=1068183458&amp;stdate=&amp;enddate=" target="_blank" class="btn">사업자정보확인</a>
-					</p>
-					<p>이메일주소 무단수집거부 : MasterPiece@mp.co.kr</p>
-					<p>
-					개인정보책임관리자 : MasterPiece 
-					｜
-					통신판매업신고 : 용산 제 012345호
-					</p>
-					<p>분쟁조정기관 : 소비자보호원, 전자거래분쟁중재위원회</p>
-					<p>호스팅 제공자 : (주)MasterPiece ｜  구매안전 서비스 이용가능
-					<a href="javascript:window_open('http://www.compuzone.co.kr/pop_page/pop_escrow.htm','535','620','scrollbars=yes','');" class="btn">자세히보기</a>
-					</p>
-				</li>
-				<li class="info_conpany">
-					<p style="font-weight:bold; font-size:14px;">고객센터</p>
-					<p>방문상담 · 구매 · A/S</p>
-					<p>역삼점 : 서울특별시 강남구 테헤란로</p>
-					<p style="margin-top:25px;"><b>택배 A/S 접수</b></p>
-					<p>서울특별시 강남구 테헤란로</p>
-					<p>ㅤ</p>
-					<p>ㅤ</p>
-				</li>
-				<li class="info_tel">
-					<p style="font-weight:bold; font-size:14px;">전화보다 빠른 상담!</p>
-					<p style="font-size:13px;">업계최초 10분 답변 서비스</p>
-					<p><a href="http://www.google.co.kr" class="btnB" style="margin:4px 0;">1:1 문의하기</a></p>
-					<p><strong style="display:inline-block; font-size:12px;">Tel 1588-0123</strong> (유료) <a href="http://google.co.kr" class="btn">전화 전 클릭</a></p>
-					<p style="color: #888888; font-size: 11px; margin-bottom: 3px; letter-spacing: -0.5px; line-height: 120%;">영업시간 9:30 ~ 18:30 (토·일·공휴일 휴무)<br>팩스번호 02-123-4567</p>
-				</li>
-			</ul>
-		</div>
-		<br><br>
-	</div>
-		<p class="footerP" align="center" style="font-size:14px;">Copyrightⓒ. All Rights Reserved </p>
-	<br><br>
-
-</body>
-</html>
